% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dftomat}
\alias{dftomat}
\title{dftomat}
\usage{
dftomat(obj, cols)
}
\arguments{
\item{obj}{\code{data.frame}}

\item{cols}{\code{character} \code{vector} of columns to convert to matrix}
}
\value{
A \code{matrix}
thedf <- data.frame(ID=rep(c('A', 'B'), length.out=10), A=sample(10), B=rnorm(10), C=sample(10))
dftomat(thedf, cols=c('A'))
dftomat(thedf, cols=c('A', 'B'))
dftomat(thedf, cols=c('A', 'B', 'C'))
}
\description{
Converts selected columns of \code{data.frame} to a \code{matrix}
}
