% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.r
\name{expand_column_values}
\alias{expand_column_values}
\title{expand_column_values}
\usage{
expand_column_values(column, values, index_i, index_j)
}
\description{
expand_column_values
}
\examples{

thedf <- tibble::tibble(
ID=rep(LETTERS[1:3], length.out=10),
x=sample(10),
y=sample(10),
extra1=sample(letters, size=10),
extra2=sample(letters, size=10),
extra3=sample(10)
)
expand_column_values('extra1', thedf$extra1, index_i=c(1, 3), index_j=c(2, 4))
}
