% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.r
\name{threshold_distance}
\alias{threshold_distance}
\title{threshold_distance}
\usage{
threshold_distance(
  data,
  threshold,
  cols = c("x", "y"),
  id_col = "ID",
  extra_columns = NULL,
  as_dataframe = FALSE,
  check_id = TRUE
)
}
\arguments{
\item{data}{\code{data.frame} of data to compute distance of}

\item{threshold}{Maximum distance to return}

\item{cols}{Names of columns of numeric data. The data will first be sorted on the first of these.}

\item{id_col}{Name of column holding ID data}

\item{extra_columns}{Names of other columns to expand into the results based on indices.
Two new elements will be made for each, one for the i index and one for the j index.}

\item{as_dataframe}{\code{logical} if a \code{list} (default) or \code{data.frame} should be returned}

\item{check_id}{Whether the ID variable should be checked for inclusion}
}
\value{
Either a \code{list} or \code{data.frame} showing which IDs matched with other
IDs, the distance between them and the rows numbers where the pairs occured.
}
\description{
Computes the distance between rows and returns those that fall below \code{threshold}
}
\details{
Computes the distance between rows and returns those that fall below \code{threshold}.
If two rows have the same ID, they will not be compared and the row-pairs will not be returned.
}
\examples{
thedf <- data.frame(
ID=rep(LETTERS[1:3], length.out=10),
x=sample(10),
y=sample(10),
extra1=sample(letters, size=10),
extra2=sample(letters, size=10),
extra3=sample(10)
)

threshold_distance(thedf, threshold=3, as_dataframe=FALSE)
threshold_distance(thedf, threshold=3, as_dataframe=TRUE)
threshold_distance(thedf, threshold=3, as_dataframe=TRUE, check_id=FALSE)
}
\author{
Jared P. Lander
}
