% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.r
\name{threshold_distance2}
\alias{threshold_distance2}
\title{threshold_distance2}
\usage{
threshold_distance2(
  left_df,
  right_df,
  threshold,
  cols = c("x", "y"),
  id_col = "ID",
  extra_columns = NULL,
  as_dataframe = FALSE,
  check_id = FALSE,
  distance_type = c("euclidean", "haversine")
)
}
\arguments{
\item{left_df}{\code{data.frame} old data to compare distances to}

\item{right_df}{\code{data.frame} new data to check for points in \code{left_df} that are nearby}

\item{threshold}{Maximum distance to return}

\item{cols}{Names of columns of numeric data. The data will first be sorted on the first of these.}

\item{id_col}{Name of column holding ID data}

\item{extra_columns}{Names of other columns to expand into the results based on indices.}

\item{as_dataframe}{\code{logical} if a \code{list} (default) or \code{data.frame} should be returned}

\item{check_id}{Whether the ID variable should be checked for inclusion}

\item{distance_type}{What distance function to use}
}
\value{
Either a \code{list} or \code{data.frame} showing which IDs matched with other
IDs, the distance between them and the rows numbers where the pairs occured.
}
\description{
Computes the distance between rows of \code{left_df} (old data)
and \code{right_df} (new data) and returns those that fall below \code{threshold}.
}
\examples{
left_df <- data.frame(
x = c(0, 1, 1, 0),
y = c(0, 1, 3, 2)
)

right_df <- data.frame(
   x = c(0, 0, 2),
   y = c(1, 3, 2)
)

threshold_distance2(left_df, right_df, threshold = 1.5, as_dataframe=FALSE)
threshold_distance2(left_df, right_df, threshold = 1.5, as_dataframe=TRUE)
}
\author{
Michael Beigelmacher
}
