% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.r
\name{threshold_distance}
\alias{threshold_distance}
\title{threshold_distance}
\usage{
threshold_distance(
  data,
  threshold,
  x_col = "x",
  y_col = "y",
  id_col = "ID",
  as_dataframe = FALSE
)
}
\arguments{
\item{data}{\code{data.frame} of data to compute distance of}

\item{threshold}{Maximum distance to return}

\item{x_col}{Name of column holding x data}

\item{y_col}{Name of column holding y data}

\item{id_col}{Name of column holding ID data}

\item{as_dataframe}{\code{logical} if a \code{list} (default) or \code{data.frame} should be returned}
}
\value{
Either a \code{list} or \code{data.frame} showing which IDs matched with other
IDs, the distance between them and the rows numbers where the pairs occured.
}
\description{
Computes the distance between rows and returns those that fall below \code{threshold}
}
\details{
Computes the distance between rows and returns those that fall below \code{threshold}.
If two rows have the same ID, they will not be compared and the row-pairs will not be returned.
}
\examples{
mydf <- tibble::tibble(
ID=rep(LETTERS[1:3], length.out=10),
x=sample(10),
y=sample(10)
)

threshold_distance(mydf, threshold=3, as_dataframe=FALSE)
threshold_distance(mydf, threshold=3, as_dataframe=TRUE)
}
